requirements.txt
Django                        5.2.1
django-filter                 25.1
djangorestframework           3.16.0
Markdown                      3.8
psycopg2-binary               2.9.10
python-dotenv                 1.1.0
